refactor(tags): extract shared ActiveFilterBar; DRY filter strip across all screens
Replace the per-screen private filter bar implementations with a single
ActiveFilterBar widget. Drop the blue container background and style the
"Clear all" link to match the tag text aesthetic. Wire ActiveFilterBar
into both GtdListScreen and InboxScreen. Replace the "New context"
ActionChip in ContextTagPickerWidget with a plain text button that
matches tag visual weight.
